NamespaceName: add metrics for level depth
Metrics are a way to get statistics about a code base and can inform the values to use in the custom properties.

The metrics can be seen by running PHPCS with the `--report=info` option.

Example output for Yoast SEO Free:
```
PHP CODE SNIFFER INFORMATION REPORT
----------------------------------------------------------------------
Nr of levels in namespace name:
        2     =>  238 ( 52.89%)
        1     =>  212 ( 47.11%)
        ------------------------
        total =>  450 (100.00%)
```