TestDoubles: add check that all double/mock classes are so named
As discussed with Herre via Slack, all classes in test double directories should have `Double` or `Mock` as part of the class name.

Originally, the `TestDoubles` sniff checked if classes which contained `Double` or `Mock` in the class name were correctly placed in a `tests/doubles` directory.

With this commit, the sniff is adjusted to also check that all classes in the `tests/doubles` directory contain `Double` or `Mock` in the class name.

Notes:
1. If additional directories have been added to the `doubles_path` property, the files in these additional directories will also be checked.
2. It is not enforced for `Double` or `Mock` to be a class name suffix, so `Mock_Some_Class` will also be accepted (at this time).

Includes unit tests covering this change.
Includes updated documentation.