Files/FileName: allow for classes which are named the same as a prefix
When a class would be called `Yoast` and `yoast` would be one of the "forbidden" prefixes, the error message would recommend for the file to be called ".php".

This fix prevents that from happening and will leave classes which are called _exactly_ the same as a prefix alone.

Includes unit test.

Includes minor tweak for error message readability.