Files/FileName: improve tests
Some tests (interfaces/traits) were using the same file names and as the data provider is an array, that meant that array values were being overwritten.

This wasn't necessarily problematic as the values for those keys were the same in both cases, so all tests were still being run correctly, but is still something which should be fixed for improved future stability.