ValidHookName: improve handling of escaped slashes in prefixes
This commit is the result of a review of the sniff for correct handling of double slashes in hook names.

When the "namespace-like" prefix type is used and the hook name is in double quotes, the "namespace separator" slash should be escaped to prevent it from accidentally escaping the **next** character.

To that end:
* Slash-escaped backslashes in prefixes will now be properly recognized and handled by the sniff and will not lead to false positives due to unrecognized prefixes.
* If the hook name is passed as a double quoted string, an extra check will be executed to verify that all backslashes are slash-escaped.
    If this is not the case, a warning will be thrown.

Includes a minor efficiency fix as the prefix is now only checked and saved to the property earlier instead of the hook name being analyzed twice for this.

Includes unit tests for the functionality and some minor adjustments to the existing unit tests.

Fixes 242

In a further iteration on this sniff, the new warning could potentially be made auto-fixable.