Add a custom "Threshold" report for use with PHPCS
This commit adds a custom report for use with PHPCS to compare the run results with "threshold" settings.

The report will look in the runtime environment for the following two environment variables:
* `YOASTCS_THRESHOLD_ERRORS`
* `YOASTCS_THRESHOLD_WARNINGS`

... and will take the values of those as the thresholds to compare the PHPCS run results against.

If the environment variables are not set, they will default to `0` for both, i.e. no errors or warnings allowed.

To use this report, run PHPCS with the following command-line argument: `--report=YoastCS\Yoast\Reports\Threshold`.

After the report has run, a global `YOASTCS_ABOVE_THRESHOLD` constant (boolean) will be available which can be used in calling scripts, to, for instance, exit with a different exit code if the thresholds were not matched.

Note: the messaging in the report - _"update the threshold in the composer.json"_ - is strongly based on existing practices in the Yoast environment.

Includes minor tweaks to the CS rules applied to the YoastCS repo itself as this code is not run within the context of WordPress and therefore doesn't have to (and often can't) comply with WordPress specific rules.