Threshold report: use colour constants
This changes the colourization of the text messages to be more descriptive by using class constants for the colour codes.

As a lot of the `echo` statements can now no longer use double quoted strings (due to the references to class constants), this also changes the `echo` statements to use comma separated arguments instead of concatenation for creating the output.