PHPCS: use PHPCompatibility proper, not PHPCompatibilityWP
The YoastCS code is not run in the context of WP, but the Yoast standard uses the `PHPCompatibilityWP` ruleset which excludes PHP features polyfilled in WP from being flagged.

By setting the `severity` of all the sniffs in the `PHPCompatibilility` ruleset back to `5`, the `exclude`s from the `PHPCompatibilityWP` ruleset are effectively "undone" and we are back to using `PHPCompatibility` proper.