Commenting/CodeCoverageIgnoreDeprecated: prevent false negatives with attributes
PHP 8.0+ attributes can be placed between a docblock and the function declaration it applies to.

The `Yoast.Commenting.CodeCoverageIgnoreDeprecated` sniff did not yet take this into account.

This would result in a false negatives as the sniff would bow out for not finding a docblock, even when the docblock contained a `@deprecated` tag.

Fixed now.

Includes unit tests.