Commenting/CoversTag: make more code-style independent
The regex used in the `Yoast.Commenting.CoversTag` sniff was - in part - based on the known naming conventions used in WP projects.

As the Yoast standard will now also be used outside of WP projects (special projects team), the assumption that the WP naming conventions will be used needs to be removed.

Fixed now.

Includes updated and additional unit tests which would not pass without this fix.