Commenting/FileComment: efficiency tweak
While reviewing the sniffs for WP-code style independence (for use in special projects), I noticed that the `FileComment` sniff may walk all the way to the end of a file before deciding to pass off to the parent sniff.

While for namespaced files, this will never happen as a namespace declaration will (normally) be the first statement in a file - with the possible exception of a `declare` statement.
For non-namespaced files, this made the sniff highly inefficient as, especially for long files, it would have to walk through all the tokens in the file before deciding it is a non-namespaced file.

Fixed now by only walking as far as the first non-empty (not whitespace, not comment), non-declare token and deciding based on that.

While a file _may_ contain a namespace declaration later on in the file (multi-namespace files), this rule is not intended for those type of files anyway, so it is correct for the "must-have file comment" rule to be applied in that case.

This change is already covered by the existing unit tests.
Includes additional unit tests specifically for the handling of the `declare` statements though.