Commenting/FileComment: minor bug fix
The PHPCS `File::findNext()` function will return the stack pointer to the token found or `false` if no matching token is found.

The check for the `$comment_start` token could therefore return `false` when no comment token is found between a PHP open tag and the first namespace token, which would mean that the `$tokens[ $comment_start ]['code']` check in the condition would effectively look at `$tokens[ false ]['code']`, which (in most cases) would translate to the PHP open tag.

While this doesn't lead to any problems in practice, it is still wrong, so a check against `false` has been added to the condition now.