Commenting/TestsHaveCoversTag: report missing covers tag when docblock is missing
The sniff as it was, contained an assumption that all methods should have a docblock, as per the WordPress Coding Standards.

For a `@covers` tag to be found, a docblock is necessary, but as-it-was, if the docblock was not found, the sniff would bow out and defer to the WPCS Docs sniff to report the missing docblock.

For the sniff to be properly independent of other sniffs, it should not bow out when there is no docblock, but report the missing `@covers` tag.

Fixed now.

Includes adjusting the expected result for one of the existing unit tests (which already covered this case).