Commenting/TestsHaveCoversTag: allow for abstract test methods in TestCases
Now, `TestCase` classes are taken into account, we also need to account for the possibility of those containing `abstract` test methods.

In that case, no `@covers` tag is needed as the method is not implemented (yet) and the `@covers` tag should only be added with the concrete implementation in the child class.

Fixed now.

Includes additional unit test.