NamingConventions/ObjectNameDepth: allow for class names in CamelCaps
The WP Coding Standards and by extension YoastCS, expects `Snake_Caps` class names.

For non-WP projects, however, `CamelCaps` class names are more common.

This change makes this sniff more code-style independent by allowing for `CamelCaps` class names when examining the word count.

Includes unit tests.

Note: one unit test is failing due to an acronym being used in the class name. This should be fixed at a later point in time.

Also note: as PHPCS has strict directory structure conventions, it is not possible for sniffs to be placed in a deeper subdirectory/namespace (which is what should normally be done for violations against this sniff).
With that in mind, this sniff is disabled for this repository.