Commenting/FileComment: allow for PHPCS disable annotation in file docblock
Previously, if a namespaced file had a deliberate docblock which shouldn't be removed, there were two options:
1. Add an `<exclude-pattern>` based selective ignore to the ruleset for the file.
2. Add a `// phpcs:ignore ...` comment ABOVE the file docblock to prevent the docblock from being reported.

This modification changes the sniff to allow for a `phpcs:disable ...` comment for this sniff **_within_** the docblock, which makes for tidier code.

It does so by searching the found file docblock for a PHPCS disable annotation disabling this particular sniff and if found, this will prevent the warning from being thrown.

Includes unit tests.