GH Actions: tweak the way the PHPCS versions are set
As things were, whenever the minimum PHPCS or WPCS version would be changed, the branch protection settings for both the `main` and the `develop` branch would need to be updated and all "required builds" referencing the old PHPCS/WPCS version would need to be removed, while new "required builds" would need to be added referencing the new minimum PHPCS/WPCS version.

This was a fiddly process and time-consuming.

The change proposed in this commit takes advantage of the Composer `--prefer-lowest` setting to achieve the same without a hard-coded PHPCS/WPCS version in the build name, which means that once the branch protection settings have been updated for this PR, they shouldn't need updating anymore for future PHPCS/WPCS version bumps.

Notes:
* Includes a new debug step ("Verify installed versions").
* The `--prefer-lowest` command is not run via the `ramsey/composer-install` action runner, but in a separate step. Running it via `ramsey/composer-install` would also impact the PHPUnit version, which would break the builds.
* I'm only making this change in the `test.yml` workflow, not in the `quicktest.yml` workflow.
    - The builds in `quicktest` are not _required_, so do not impact the branch protection settings.
    - This also allows to still test mixed high/low variants (i.e. low PHPCS, high WPCS and visa versa) via the `quicktest` workflow.