Drop support for PHP < 7.2
The majority of the Yoast codebases now have a minimum of PHP 7.2.

Additionally, the YoastCS checks should give the same results independently of the PHP version on which the CS check is being run, which means that generally speaking, the CS check is run on a high PHP version, both locally as well as in CI.

With this in mind, there is no reason to keep the YoastCS minimum supported PHP version at PHP 5.4, so let's bring it in line with the other Yoast codebases and let it have a PHP 7.2 minimum.

Note: PHP_CodeSniffer itself still has a PHP 5.4 minimum. PHPCS 4.x is expected to have a PHP 7.2 minimum and this change is in line with that.