Docs: remove redundant `@package` tags
`@package` tags are an arcane manner to group related files as belonging to one project.

For projects using namespaces, the current recommendation is to only have `@package` tags when they have supplemental information to the namespace.

As that isn't the case in YoastCS (with the exception of the non-namespaced test bootstrap file), I'm proposing to remove the `@package` tags from the YoastCS class docblocks.

Along the same lines, the `@author` tag is also an outdated practice as `git blame` will work just fine, so I'm removing those too.

Includes cleaning up (normalizing) the tag description alignments in the class docblocks.

:point_right: reviewing with whitespace changes ignored should make it easier to see that the only real change is the removal of the package/author tags.

Refs:
* https://docs.phpdoc.org/3.0/guide/references/phpdoc/tags/package.html#package
* https://github.com/php-fig/fig-standards/blob/master/proposed/phpdoc-tags.md#59-package