NamingConventions/ValidHookName::transform(): sync parameter names
The parameter name of the upstream method was updated in WPCS 3.0.0 to not overlap with a reserved keyword.
And for potential future support of function calls to sniff methods using named parameters, the signature of a method in a child class overloading a method in a parent class should be kept in sync.

This makes it so for the `NamingConventions/ValidHookName::transform()` method.