GH Actions/test: remove unused steps for PHPCS 4.x
Testing against PHPCS 4.x-dev has not been possible for a while and development of that version looks to be stagnated anyway, so testing against it is not really relevant at this time.

This commit removes all references to testing against PHPCS 4.x from the workflow.

If/when testing against the next PHPCS major becomes relevant again, this can brought back and adjusted based on the reality at that time.