YoastCS/Modernize - Tests: add PHP native return type declarations
As support for PHP < 7.2 has now been dropped, PHP native parameter and return type declarations can be used.

There are, however, a couple of caveats:
1. Return type declarations can be freely used as long as the type needed is available in PHP 7.2.
2. Parameter type declarations can **_not_** be freely used as parameter types are contravariant, which means they can only be _wider_ than the original type, which means parameter type declarations can only be added for YoastCS native methods and not for methods which implement a PHPCS native interface method or overload a PHPCS native sniff method, as in that case, the original method will rarely have a type declaration, which means that adding one in YoastCS would violate LSP.

Also note that a `TypeError` thrown by a sniff, would stop the scan of a file dead, so let's be conservative with adding type declarations in the sniffs for now, but add them freely to the tests.