Commenting/CodeCoverageIgnoreDeprecated: minor code simplifications
Remove two redundant conditions.

The first is not needed as it's not actually important what token type we encounter, the only thing important is that there is a known attribute opener to skip to.

The second is not needed as the `findPrevious()` call will always find something, if nothing else, it will find the docblock opener, so checking for `false` is unnecessary.