Commenting/CoversTag: flag deprecated @covers tag formats
PHPUnit 9.0 (soft) deprecated the use of `ClassName<*>` type `@covers` annotations. Support for these type of annotation has been removed completely in PHPUnit 10.0.0.

This commit adds a new _warning_ to the sniff, which will detect the use of these deprecated formats and flag them.

The choice for _warning_ vs _error_ is deliberate. The Yoast repos, for now, do not use PHPUnit 10.x yet, so for now, a warning is sufficient.
Once the repos would be upgraded to start using PHPUnit 10.x, this sniff should be updated and the `warning` changed to an `error`.

Includes tests.
Includes a new section in the XML docs about this new warning.

Refs:
* sebastianbergmann/phpunit 3630 (PHPUnit 9.0 deprecation).
* sebastianbergmann/phpunit 3631 (PHPUnit 10.0 removal).