Yoast/AlternativeFunctions: bug fix - fixer could create parse error
If the original function call was already fully qualified in a namespaced file, the fixer would add a superfluous extra namespace separator - `\\WPSEO_Utils::format_json_encode()`, effectively creating a parse error in the resulting file.

As fully qualified calls to global functions in namespaced files are now the norm in the Yoast repos, this is a clear bug which should be prevented.

Fixed now by removing a potential leading namespace separator before doing the replacement.

Includes additional tests.