Files/TestDoubles: implement use of new PathHelper and PathValidationHelper classes and check case-sensitively
Aside from starting to use the new `PathHelper` and `PathValidationHelper` utility function classes, this commit also changes the path comparison from case-INsensitive to case-sensitive, which, what with the change to strict PSR-4 compliance for all test files, including double/mock files, is a change which needs to be made anyway.

The tests sub-directory names have been updated to proper case to be in line with this change, the inline property settings in test case files have been updated too.

On top of that, a couple of dedicated new test cases have been added to verify the case-sensitive handling of the `double_path` property.

Includes updating two pre-existing tests to pass duplicate excluded files in different ways.