NamingConventions/NamespaceName: improve path handling
This changes the `src_directory` property handling.

Originally, the file paths would always be stripped down to a relative path in relation to the `basepath`.
Now, the `src_directory` path comparison(s) will always be based on the absolute path, including the `basepath`.

This should have a small performance benefit, in that some of the path manipulation is moved to the `validate_src_directory()` method, the logic of which under normal circumstances, will only be run once, while the `process()` method is run for every `namespace` keyword encountered.

This also implements the use of some additional functions from the `PathHelper` and the `PathValidationHelper` classes.

Includes updating a pre-existing test to pass duplicate excluded files in different ways.