Files/TestDoubles: remove "one object per file" check
This is already checked via the `Generic.Files.OneObjectStructurePerFile` sniff, which is included in the WordPress Coding Standards. No need for doubling this check here.

Includes removing the associated unit tests.
Includes removing the code sample related to mocks/doubles being in the same file as the test class from the XML docs.