Files/TestDoubles: make the sniff codebase agnostic
The `public $doubles_path` contained an arbitrary default value.

While this default value is applicable for some of the Yoast codebases, it still needs to be duplicated in the ruleset due to a bug/missing feature in PHPCS in how array properties containing default values are handled.

Having this arbitrary default value in the sniff also codifies some Yoast-specific repo layout logic in the sniff, while sniffs, generally speaking should be code-base agnostic.

With that in mind, I'm removing the default value. The `Yoast` ruleset contains this value anyhow (due to the bug/missing feature) and that is the more appropriate place for code-base specific/organisation specific property values.

Includes updating the test case files to allow for this change.

Refs:
* squizlabs/PHP_CodeSniffer 2154
* squizlabs/PHP_CodeSniffer 2228