Files/FileName: efficiency fix
Always return a stack pointer at the end of the file, even when the file passed comes from `STDIN`, as we ignore `STDIN` files for this sniff, so there is no need to continue listening for tokens in the file.

Includes ignoring this line for code coverage checks as the line is untestable in a unit test situation.