Files/FileName: bug fix - excluded paths were not always normalized correctly
The `is_file_excluded()` method would first call the `clean_custom_array_property()` method with the `$flip` parameter set to `true`, resulting in an array containing the excluded files as the keys.

After that it would call the `normalize_directory_separators()` method on all _values_ via an `array_map()`.

In effect, this means that the actual values, which are stored in the array _keys_ would still not be normalized, meaning the "is file excluded" check could return the wrong result for ruleset provided paths containing a backslash directory separator.

Even though, in practice, this didn't lead to problems as the paths in rulesets are set using forward slashes, it is still a bug.

Fixed now by no longer "flipping" the array.

And as the `$flip` parameter of the `clean_custom_array_property()` method is now no longer used, the parameter and its handling has been removed from the function.

Includes test.