Files/FileName: only clean up the excluded files when needed
Efficiency tweak. Cleaning the ruleset passed "excluded files" doesn't need to be done time and again every single time this sniff is called.

For a normal PHPCS run, where the property is set in a ruleset, doing it once and reusing the cleaned up version in subsequent calls to the sniff will be sufficient.

For test runs, this may need to be done more often, but that can be handled by storing the previous value of `$excluded_files_strict_check` property and checking if it has changed before doing the validation.

This commit implements this.

Additionally it adds some extra safeguards for incorrectly passed "excluded files" paths.

Includes removing the `$phpcsFile` parameter from the `is_file_excluded()` method as it is no longer needed by that method.

Includes additional unit tests.
Includes updating a pre-existing test to pass duplicate excluded files in different ways.