Utils/CustomPrefixes: make methods `final`
Traits cannot be `final`, but their methods can be, though this will only have an effect if a class `use`-ing the trait is being extended.

All the same, the `final` keyword on the methods should serve as a warning to anyone who wants to overload these methods that they probably shouldn't.