NamingConventions/ObjectNameDepth: bug fix for test classes using CamelCaps
Test classes and Double/Mock classes are allowed one extra word in the name to account for the `Test`/`Mock`/`Double`.

However, this was not handled correctly for test classes in CamelCaps as those are first converted to `snake_case`, which means that the suffix used in the comparison is now lowercase, while the name comparison was done in a case-sensitive manner.

Along the same lines, underscore separated class names where the suffix would not be "proper case" would also not be recognized correctly as, again, the name comparison was done in a case-sensitive manner.

As the casing of the class name is not the concern of this sniff, the name comparison should be done in a case-INsensitive manner.

Fixed now.

Includes tests.