NamingConventions/ObjectNameDepth: remove some redundant code
The WPCS `IsUnitTestTrait::is_test_class()` method does not (and never has so far) handle import `use` statements, which means that in the context of namespaced files, the results of the function are unreliable.

As documented with the `TEST_SUFFIXES` constant, the `Test` and `TestCase` suffixes should only be given the "one extra word" treatment when these classes extend a "known" test class.

As things were, this wasn't respected properly by the sniff anyhow (the `$this->is_test_class()` should have been _nested_ within the `self::TEST_SUFFIXES[ $last ] === true` condition).

And as it cannot reliably be enforced at this time, we're better off removing the useless condition and just checking if a class with one of the "extra word" suffixes extends.