NamingConventions/ValidHookName::verify_yoast_hook_name(): change function signature
Originally, the `verify_yoast_hook_name()` method would get passed the complete stack of parameters.

This is not necessary as the method only examines the first - `$hook_name` - parameter and that parameter has already been retrieved from the stack before.

This changes the function signature of the `verify_yoast_hook_name()` method to only expect the parameter information for the `$hook_name` parameter, instead of the complete stack of parameters.

It also changes the method from `public` to `private`. This method never needed to be `public`.