NamingConventions/ValidHookName: add support for PHP 8.0+ named parameters
While WP does not officially support function calls using named parameters, that doesn't mean that nobody will use them....

This adjusts the sniff to allow for calls to hook functions using named parameters.

The way the correct parameter is retrieved now uses the new `WPHookHelper::get_hook_name_param()` method, which uses the PHPCSUtils `PassedParameters::getParameterFromStack()` method under the hood. The `WPHookHelper` class keeps track of which functions are considered hook functions, what parameter names are used in WP and the parameter position.

Note: as parameters can be skipped and/or passed in an unconventional order, it is now possible for the parameter to not exist. While that would result in a fatal error in PHP (missing required parameter), the sniff should still handle this situation gracefully, which is why the check for the parameter has been moved up and will cause the sniff to bow out if the parameter can not be found.

Includes additional unit tests.

Note: function calls with named arguments are not supported for the `do_action()` or `apply_filters()` functions as those have a parameter with a spread operator (but that's irrelevant for the sniff).