NamingConventions/ValidHookName: add some extra defensive coding
As things are, we already know that `$first_non_empty` will always return an integer stack pointer, as the same check is done in the `process()` method and this method isn't called if the return would have been `false`.

Having said that, that means these methods are (too) closely coupled and a change in the logic in the `process()` method could inadvertently cause PHP to start throwing notices in the `verify_yoast_hook_name()` method.

So, I'm adding some extra defensive coding just in case, though also ignoring that code for code coverage as, as things are, the condition will always be `false`.