NamingConventions/NamespaceName: improve handling of unconventional directory names
As things were, the `NamespaceName` sniff would translate the namespace to the expected directory path and then compare the directory paths.

This is the wrong way round as this sniff is about namespace names, not directory names, so the directory name should be translated to a namespace name and then compared with the actual namespace name instead.

Along the same lines, there was the possibility that the sniff would recommend an "Expected" namespace name which would be a parse error/invalid syntax in PHP.

This commit fixes both these issues.

The commit also adds an extra error for directory names which can't (for a certain definition of "can't") be translated to a valid namespace name.

Includes additional unit tests demonstrating the problem.

Note: if the name in use in the file is causing a problematic parse error (like in the test with the `#` in the namespace name), the sniff will stay silent. This is the normal behaviour for PHPCS check when encountering parse errors.