NamingConventions/NamespaceName: throw an error if a prefix is expected, but not used
The sniff expects a `prefixes` property to be passed from a ruleset and for a valid prefix to be used in a namespace name, however, this was not checked.

This commit adds a new error to the sniff which will be thrown if at least one valid prefix has been passed and no valid prefix was used in the namespace name.

Includes unit tests + some minor updates to pre-existing tests which will now also throw this error.