NamingConventions/NamespaceName: make the "incorrect namespace name" error more informative
When no prefix was matched, the error would previously read `Expected: "[Plugin\Prefix]..."; Found: "..."`.
However, if there is only one known valid prefix, we can already replace the `[Plugin\Prefix]` with the expected prefix.

Fixed now.

The effect of this fix can be seen in the messages thrown by the unit tests, but can't be safeguarded via the tests at this time.